Add path validation when deserializing
